payment fix 001

+++ apps/store/api/order.py
@@ -251,12 +251,12 @@ class CustomerFunctions(ViewSet):
         )
 
     def to_review_items(self, request):
+
         customer = get_customer(request.user)
         order_items = OrderItems.objects.filter(
             order__customer=customer, has_review=False
         )
         data = []
-
         for row in order_items:
             data.append(
                 {
@@ -277,7 +277,6 @@ class CustomerFunctions(ViewSet):
 
         if data:
             return Response(data={"reviews": data})
-
         return Response(data={"reviews": None})
 
 
@@ -290,39 +289,40 @@ from django.views.decorators.csrf import csrf_exempt
 def order_payment_confirm_webhook(request):
     payload = request.body
     event = None
-    sig_header = request.META["STRIPE_SIGNATURE"]
-
-    print(f"sig_header {sig_header},\n endpoint_secret {endpoint_secret}")
-    try:
-        event = stripe.Webhook.construct_event(payload, sig_header, endpoint_secret)
-        if event.type == "checkout.session.completed":
-
-            order_id = event["data"]["object"]["metadata"]["order_id"]
-            try:
-                order_queryset = Order.objects.get(id=order_id)
-                order_queryset.payment_status = True
-                order_queryset.save()
-                cart = Cart.objects.filter(customer=order_queryset.customer).first()
-                if cart:
-                    CartItem.objects.filter(cart=cart).delete()
-            except Order.DoesNotExist:
-                ...
-
-        else:
-            print("Unhandled event type {}".format(event.type))
-
-            if event.type == "checkout.session.async_payment_failed":
+    # print(request.META)
+    sig_header = request.META.get("HTTP_STRIPE_SIGNATURE")
+    if sig_header:
+        print(f"sig_header {sig_header},\n endpoint_secret {endpoint_secret}")
+        try:
+            event = stripe.Webhook.construct_event(payload, sig_header, endpoint_secret)
+            if event.type == "checkout.session.completed":
                 order_id = event["data"]["object"]["metadata"]["order_id"]
                 try:
                     order_queryset = Order.objects.get(id=order_id)
-                    order_queryset.delete()
+                    order_queryset.payment_status = True
+                    order_queryset.save()
+                    cart = Cart.objects.filter(customer=order_queryset.customer).first()
+                    if cart:
+                        CartItem.objects.filter(cart=cart).delete()
                 except Order.DoesNotExist:
                     ...
 
-            print("Unhandled event type {}".format(event.type))
-    except Exception as e:
+            else:
+                print("Unhandled event type {}".format(event.type))
 
-        print(e)
-        return HttpResponse(status=400)
+                if event.type == "checkout.session.async_payment_failed":
+                    order_id = event["data"]["object"]["metadata"]["order_id"]
+                    try:
+                        order_queryset = Order.objects.get(id=order_id)
+                        order_queryset.delete()
+                    except Order.DoesNotExist:
+                        ...
 
-    return HttpResponse(status=200)
+                print("Unhandled event type {}".format(event.type))
+        except Exception as e:
+            print(e)
+            return HttpResponse(status=400)
+
+        return HttpResponse(status=200)
+
+    return HttpResponse(status=400)
